pyproject.toml
[tool.poetry]
name = "litellm-study"
version = "0.1.0"
description = ""
authors = ["Serena <younginsong.eng@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
litellm = {extras = ["proxy"], version = "^1.57.8"}
openai = "^1.59.6"
pydantic = "^2.10.5"
python-dotenv = "^1.0.1"
loguru = "^0.7.3"


[tool.poetry.group.dev.dependencies]
ruff = "^0.9.1"
black = "^24.10.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
